refactor(shared): organize Entity and IRepository into Entities and Repositories folders
